handle.fd: fix bug with not correctly removing closed fds
We left the empty list in fd_table_to_near_to_handles. This means that
garbage collection would later try to close this fd - even though we'd
already closed it!

Luckily, if rsyscall is the only thing opening file descriptors, the
list will be reused and new file descriptors with the same number will
have their handles stored in that same list.

But there's still the possibility of closing something that is already
closed and getting EBADF. I'm not sure why we never hit that, except
perhaps that it's because we hit the highest number of open file
descriptors right before calling unshare_files, which is when GC
happens.

If rsyscall isn't the only thing opening file descriptors, there's
also the possibility of closing something that belongs to someone else
in the process. (That's how I found this)