handle.fd: add ability to inherit fds after-the-fact
Previously inheritance was tied to a task; to inherit an fd handle
into a newly-copied fd table, we needed to have a handle for that fd
in the task that was changing fd table. But that doesn't work for the
case of clone, which creates a newly-copied fd table and a new task at
the same time.

We resolve this by making it explicit that a new fd table starts out
with a copy of all the fds in its parent; and reserve the ability to
"inherit" an fd handle into the new fd table by checking if that
handle was present in our parent at the time the new fd table was
copied.

This is a bit ad-hoc at the moment, but the fundamental principle is
very sound, I think.