tasks.ssh: disable privsep for local testing sshd
Old versions of openssh (6.7 at least) have a bug where they check for
the presence of the privsep directory, even if already running as
non-root and therefore not using privsep. The only recourse is to
disable privsep entirely.

This doesn't really matter because, as previously stated, we already
weren't using privsep since we're running as non-root.