clone: initialize child_ctid futex to non-zero to remove race
The kernel zeros child_ctid and does a FUTEX_WAKEUP on it, once the
child has exited or exec'd.

If we leave child_ctid with an initial value of 0, and do a
futex(futex, 0, FUTEX_WAIT), then there's a race condition: If the
child exits before we start waiting, then the kernel will do the
wakeup before we start waiting, and once we wait we will never wake
up.

This is a fairly basic race condition. I think we missed this because
we confused the child_ctid futex with the robust list futexes; the
robust list futexes have a non-zero value written to them before being
woken up, so an initial value of zero is correct for them.

We started seeing this race condition be hit in production on some
heavily loaded machines when constrained to a single CPU.