clone: remove unshare flag, make no CLONE.FILES the default
We prviously defaulted to having CLONE.FILES|CLONE.IO|CLONE.SYSVSEM
set on calls to clone, and they could be unset by passing those flags
in the unshare argument. Now, instead, we default to CLONE.NONE.
That's closer to the real interface of clone.

We still force CLONE.VM to be set, since that's required for
our exec-detection approach.