FileDescriptor: add __int__ method so int(fd) works
Having to do .near is unnecessary complication.

I was concerned that people will convert file descriptors to integers
in places where it's not necessary, but in practice I don't think that
will happen, because there aren't any public rsyscall functions that
take integer file descriptors, and all the syscalls they might want to
call are available as methods.