fd: support inheritance of implicitly-inherited fds
If a task clones, then clones again, the grandchild will have
inherited its fds, because it inherits everything the child had. But
we didn't have the bookkeeping for that previously; instead the
grandchild would seem to have inherited nothing.

Now we include inherited fds in the list of inherited fds, so when
inheritance is iterated without intervening calls to inherit_fd, it
will work correctly.