remove some pointer borrows
Borrowing these pointers is not useful right now, and slows things
down quite a bit, since ExitStack is apparently very slow.

The long term use of this is that, if we call a function to pin the
pointer while we're using it, we could have our memory allocator move
pointer allocations transparently (GC-style) and thereby get more
efficient.  That would be kind of complex, though, so we're unlikely
to do it soon. If we ever seriously want to do that, we can add the
infrastructure back in at that time.

These borrows also could be used to prevent freeing or writing to the
pointers during syscalls. But 1. they aren't checking that right now,
and 2. a better way is to just disallow manual freeing and let the
Python GC handle freeing, and let the user write if they want - this
isn't Rust, sadly.