local: insert checkpoints in local sysif and transport
These are necessary if we're relying on a task to be cancelled by
another task. We can easily write a function which only calls local
operations. Such a function previously never had any checkpoints at
which it could be cancelled, nor would other tasks ever run (so they
couldn't cancel it either).