connection: return working channels from open_channels(n>1)
make_n_in_parallel is in hindsight obviously buggy here - if we
connect and accept in parallel, we'll get channels which connect to
each other, instead of to the FDs they are paired with.

This slows things down a fair bit. It would be better to pipeline the
channel creation. But we don't have a good way right now to pipeline
syscalls while mantaining their order.

We can revisit it once we've reworked the syscall interface to provide
FIFO semantics - with those semantics, it might actually be correct to
use make_n_in_parallel.