add __hash__ to some dataclasses by setting frozen=True/eq=False
The goal is to be able to use these types as keys into dictionaries.

We make some types frozen, which makes dataclass generate an __eq__
method which compares them by their fields, and also a __hash__ which
just hashes the fields.

For Task, we disable __eq__, because Tasks should be compared by id,
not by their fields. But, that just makes Task have an id-based
__hash__, which works fine.