monitor: don't use child's ram when inheriting to child
Using the child's RAM is clearly wrong: We need a RAM for reading from
the sigfd, and for calling waitid, and both of those happen in the
parent, and should use the parent's RAM, which we already have.

This manifested itself as failures when a child creates a grandchild
and then changes address space, such as through exec. From the
perspective of the monitor, that's fine, since the grandchild is
monitored by the parent, and with this change, it works.