Pointer: store type of pointed-to memory at runtime
Previously, we only had the serializer, since the only runtime
operations we actually need are serialization and deserialization.
That works fine.

But, it's not particularly idiomatic for Python, where we would expect
be able to introspect a Pointer and learn its type.

So now we store the type.  For now, we just use it to make a nicer
stringification of the Pointer; but it's also use for building a UI
with introspection.

We still need the serializer, since it might have some extra data
associated with it.  Later on, we might be able to restructure things
later to rely more on the type, and maybe sometimes not pass the
serializer, or pass a generic serializer that takes the type.  Or
maybe not; the serializer is a pretty elegant implementation of ad-hoc
polymorphsim. (It's directly inspired by dictionary-passing typeclass
implementations)