handle.pointer: add {Readable,Linear}Pointer classes
This allows us to encode more information in the type system, but
primarily gives us some more dynamic type information, that can be
used to e.g. dynamically enforce linear reading of LinearPointer.

This dynamic information can also be used for introspection for
building a UI.

We aren't yet actually using this information in syscall return
values. That requires a lot of effort...