signal: check explicitly blocked signals ⊆ signal mask, not =
Previously, we checked for complete equality. This guarantees that no
other component in the process is allowed to change the signal mask.

But, we don't really care about that; we'd be fine with allowing
others to add blocks to the signal mask, we just don't want them to
unblock our signals.

We also want to allow signal blocks to be inherited from the
non-rsyscall parent of our local process.  (Note that we explicitly
track inheritance between rsyscall parents and children.) The signal
mask is not reset on execve, so a parent process can block signals,
which will stay blocked for its children.

But with our old model, a non-rsyscall parent could not set a signal
block and pass it down to a child using rsyscall; rsyscall would raise
an exception due to the mismatch between the expected and actual
signal mask.

The JVM, in particular, has a bug where it leaks a block for SIGQUIT
to children, which triggers this:
https://bugs.openjdk.java.net/browse/JDK-6624838

With the new model, we're tolerant of signal blocks set by other
components in the same process, and of signal blocks inherited down
from the parent.