tasks.exec: make a new sycall connection from use after exec
Previously, we would reuse the old pair of connected sockets for
sending syscalls after rsysall_exec. This was a slight optimization,
since we wouldn't need to connect some new sockets.

Furthermore, we had to set up a robust_list futex for the process
after rsyscall_exec. This was because we needed to be able to detect
the success of exit or exec without relying on EOF, since EOF meant
the connection was no longer usable, and we wanted to reuse the
connection if we called rsyscall_exec multiple times.

Now, we make a new pair of connected sockets for use after exec. After
exec, we use NonChildSyscallInterface, and rely on EOF to tell us if
later exits or execs are successful. Since the process is now in its
own file descriptor table, we can rely on the socket to be closed when
it exits, so relying on EOF is safe. Since we can rely on EOF, we no
longer need to create a robust_list futex, or rely on the process
being our child at all.