clone: stop monitoring the child process in ChildSyscallInterface
We don't actually need to monitor it. We only need the futex_process.

Why didn't we notice this earlier? Well, we used to have the
futex_process be optional for rsyscall_exec, so we could make syscalls
in the post-exec process to set up the new futex_process, since the
old futex_process would die when we exec. But now rsyscall_exec
doesn't even use ChildSyscallInterface, so that's not necessary.

squash! clone: stop monitoring the child process in ChildSyscallInterface