switch to using kiselio
This is a much larger commit than I would normally do, but I didn't
want to provide support for calling trio abstractions from kiselio
code, so I had to migrate the entire base system at once. Ideally we'd
just migrate the syscall connection, but that depends on:
- epoll
- memory allocation
- memory transport
- child process monitoring (for the futex process)
And those, of course, depend on the syscall connection. So they all
had to go at once.

We also had to update the local task, to make use of the trio system
task with TrioSystemWaitReadable.

And we had to update the persistent task, because with epoll_wait
continually running, we couldn't register the new activity fd like we
did in the old style - so we made Epoller._run automatically register
the activity fd.

We took the opportunity to get rid of SyscallResponse, since it
becomes fairly useless with kiselio.

We deleted MultiplexedEvent from concurrency because it's no longer
used as of this commit; the rest will be deleted later.