fuse: add FUSE_MIN_READ_BUFFER
As of around 5.4 (specifically 1fb027d7596464), the kernel now
enforces that the buffer to read from the /dev/fuse FD is of some
minimum size (8192 in practice). This is not a regression since a FUSE
daemon passing a too-small read buffer (like we did) would have run
into errors eventually anyway when userspace performed large writes.