tasks.exec: delete
This has had many unenlightening issues over the years, and it's not
actually useful. It's *maybe* an interesting demonstration of what you
can do with rsyscall - but it requires touching too many internals in
too awkward of a way. If we ever actually want to do this, we should
do it in such a way that it consumes the old Thread and creates a new
Thread, instead of preserving the same Thread so that it keeps
working.