handle: don't define constructor for FileDescriptor
__slots__ doesn't allow us to default valid to True, so previously we
set it in the constructor. But we only construct FileDescriptors in
one place anyway, so we can just pass True there instead.

We necessarily have to get rid of the constructor definitions in the
derived classes as well. Those constructors also narrow the type of
the `task' field through the type annotation on their argument. We can
achieve the same thing by just adding the field type hint to the
class body; looks like that doesn't duplicate the field, I guess
dataclasses dedups fields by name or something.