mktemp: have TemporaryDirectory inherit from Path
An alternative to this could be making a PathWithThread class, and
having TemporaryDirectory inherit from that. I could even put the
cleanup method on that class.

But I'm not doing that, because I want to keep paths as pure data,
pure identifiers, with no ability to perform operations without other
capabilities. Because ownership semantics are tricky with paths, which
makes it difficult to go all the way and turn them into proper caps.