nix: remove Store class, require nix-store to be on PATH
This is awkwardly dynamic; it seems like it would be nicer to have it
be statically known that we have the Nix binaries available.

But it's vastly easier to use, because otherwise we need to pass
around a Store along with a Thread everywhere. And that's tricky to
do, and of questionable value.

And it's not too much less flexible: if and when we need to operate in
an environment that doesn't already have Nix deployed and on
PATH (such as in enter_nix_container), we can just add it to PATH
ourselves, and which('nix-store') will work fine. This is, again,
awkwardly dynamic, but I think it's fine...