sys.capability: add new CAPs
Every time we upgrade to a version of Linux with new capabilities,
running capget as a privileged process will fail because we'll try to
convert an int (for the new capabilities) to an IntEnum for which that
int is not in range.  We probably should tolerate out of range CAPs
somehow... but this is easy for now.