Revert "linux.fuse: support fuse 7.33"
This reverts commit b6cc6356fbdc29ac420cd16ea43d5d501787350f.

This commit changed the names of fields in FUSE structs to match 7.33,
but that makes it impossible to compile rsyscall against any FUSE
headers before 7.33.  For new defines, we could work around this by
manually adding the defines, but there's no easy solution for headers.
Since we don't actually need 7.33 support, let's just revert this.