Revert "linux.fuse: support fuse 7.32"
This reverts commit 1e9afe6f012382502a1644ae788c012a7534df27.

This commit changed the names of fields in FUSE structs to match 7.32,
but that makes it impossible to compile rsyscall against any FUSE
headers before 7.32.  For new defines, we could work around this by
manually adding the defines, but there's no easy solution for headers.
Since we don't actually need 7.32 support, let's just revert this.