thread: add Process.fork() helper method
This is just a better name that gives better intuition.

We used to have a fork() method but it defaulted to passing clone
flags such that everything was shared with the parent process.  That's
not what fork actually does, so we renamed it to clone.  But then we
started defaulting to pass no (well, minimal) clone flags by default,
so now it matches what fork does, so we can actually call it fork.