persistent: rework PersistentSyscallConnection with more dneio
Basically we're replacing the "carefully" coded direct access to
self.conn with a coroutine which guards the connection and causes
access to it to block when it's not available.  I think this is
somewhat easier to understand and read?  And also, it will be much
easier to add the new write/read SyscallInterface methods with this
design, since they'll need an exact copy of the logic in
PersistentSyscallConnection.syscall.