add ability to write memory directly to the sysif
There are a few changes here which needed to be done together:

- tasks.connection has the largest change, to support writing memory
through a SyscallConnection without a separate data socket by just
writing and reading directly to the syscall fd
- tasks.persistent has a corresponding change so that
PersistentSyscallConnection works properly
- We stop using SocketMemoryTransport in favor of TaskTransport which
just takes a Task and uses the sysif inside to do memory access.

We will get rid of MemoryTransport entirely in a subsequent commit.