memory.ram: merge RAM into Task
We already baked memory access (reading/writing) into the
SyscallInterface, so it's already available wherever we have a Task.
The only other thing in RAM is the allocator, so the real meat in this
change is putting the allocator on Task.

That means we're assuming we'll have (or be able to create) an
allocator in *any* environment.  I think this is reasonable.  Also,
many syscalls practically require that you be able to allocate
memory (even if on the stack) to call them.  So it's something Linux
is already requiring for us.